Vehiculos

+++ database/migrations/2023_02_23_073827_create_kilometrajes_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
     {
         Schema::create('kilometrajes', function (Blueprint $table) {
             $table->id();
-            $table->double('kilometraje');
+            $table->integer('kilometraje');
             $table->date('fecha');
             $table->foreignId('vehiculo_id')->constrained()->on('vehiculos')->cascadeOnDelete();
             $table->timestamps();
